+++ domain/src/main/java/tachiyomi/domain/release/interactor/GetApplicationRelease.kt
@@ -19,8 +19,8 @@ class GetApplicationRelease(
     suspend fun await(arguments: Arguments): Result {
         val now = Instant.now()
 
-        // Limit checks to once every 3 days at most
-        val nextCheckTime = Instant.ofEpochMilli(lastChecked.get()).plus(3, ChronoUnit.DAYS)
+        // Limit automatic checks to once a day. Manual checks can bypass this interval.
+        val nextCheckTime = Instant.ofEpochMilli(lastChecked.get()).plus(1, ChronoUnit.DAYS)
         if (!arguments.forceCheck && now.isBefore(nextCheckTime)) {
             return Result.NoNewUpdate
         }